requirements.txt
pandas
numpy
matplotlib
statsmodels.api
statsmodels.tsa.seasonal